add user_exists and fix command_start
